requirements.txt
fastapi==0.92.0
uvicorn==0.17.6
aiofiles==23.1.0
python-dotenv==0.21.0